AbstractDependencyTracker for shared code
